requirements.txt
numpy>=1.25
scipy>=1.10
scikit-image>=0.21
scikit-learn>=1.3
torch>=2.1
torchvision>=0.16
transformers>=4.37
diffusers>=0.22
pillow>=10.0
opencv-python>=4.8
#pydensecrf>=0.7,<1
requests>=2.31
tqdm>=4.66
pyyaml>=6.0
pandas>=2.1
